+++ curso_mc/src/main/java/br/com/nelioalves/cursomc/curso_mc/resources/PedidoResource.java
@@ -26,7 +26,7 @@ public class PedidoResource {
 	private PedidoService service;
 
 	@RequestMapping(value = "/{id}", method = RequestMethod.GET)
-	public ResponseEntity<Pedido> find(@PathVariable Integer id) throws ObjectNotFoundException {
+	public ResponseEntity<Pedido> find(@PathVariable Long id) throws ObjectNotFoundException {
 		return ResponseEntity.ok().body(service.buscaPorId(id));
 
 	}
@@ -38,25 +38,5 @@ public class PedidoResource {
 				.path("/{id}").buildAndExpand(obj.getId()).toUri();
 		return ResponseEntity.created(uri).build();
 	}
-	
-//
-//	@RequestMapping(method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE)
-//	public ResponseEntity<?> save(@RequestBody Pedido Pedido) throws ObjectNotFoundException {
-//		Pedido = service.cadastrar(Pedido);
-//		return new ResponseEntity<>(Pedido, HttpStatus.CREATED);
-//	}
-//
-//	@RequestMapping(method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
-//	public ResponseEntity<Collection<Pedido>> findAll() {
-//		Collection<Pedido> PedidoCollection = service.buscarTodos();
-//		return ResponseEntity.ok(PedidoCollection);
-//	}
-//
-//	@RequestMapping(method = RequestMethod.PUT, consumes = MediaType.APPLICATION_JSON_VALUE)
-//	public ResponseEntity<Pedido> alteraPedido(@RequestBody Pedido Pedido) {
-//		// fazer a verificacao para ver se a cateforia existe
-//		Pedido PedidoReturn = service.alterar(Pedido);
-//		return new ResponseEntity<>(PedidoReturn, HttpStatus.CREATED);
-//	}
 
 }
